api/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
pydantic>=2,<3
azure-storage-blob>=12,<13
python-dotenv>=1,<2
